requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
asyncpg
alembic
aiohttp
aiomysql
async-timeout
bcrypt
passlib
pydantic
PyMySQL
SQLAlchemy
starlette
requests
pydantic[email]
cryptography
PyJWT
python-multipart
pydantic-settings